requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0.3
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-modeltranslation==0.14.4
django-recaptcha3==0.1.0
h11==0.9.0
httptools==0.0.13
idna==2.9
Pillow==7.0.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
